+++ src/app/Ambiente/ambientes/ambientes.component.ts
@@ -48,62 +48,7 @@ export class AmbientesComponent implements OnInit {
   }
 
   AbrirDialog(ambienteId:any,ambiente:any){
-    this.dialog.open(DialogLiberarAmbientComponent,{
-      data:{
-        ambienteId:ambienteId,
-        ambiente:ambiente
-      }
-    }).afterClosed().subscribe(resultado => {
-      if(resultado === true)
-      {
-        this.ambiente.GetAmbientesBackEnd().subscribe((dados) =>
-        {
-          this.dataSource.data = dados;
-          this.displayedColumns = this.ExibirColunas();
-          this.router.navigate(['/ambientes']);
-        });
-      }
-    });
-  }
-}
-
-@Component({
-  selector: 'app-dialog-liberar-ambiente',
-  templateUrl: 'dialog-liberar-ambiente.component.html',
-})
-export class DialogLiberarAmbientComponent {
-  constructor( @Inject (MAT_DIALOG_DATA) public data: any ,
-              private service : AmbienteService,
-              private snackBar : MatSnackBar) {}
-
-    ambienteLiberado!:any;
-    erros!:string [];
-
-    LiberarAmbiente(ambienteId:any,ambiente:any):void{
-    console.log(ambienteId,ambiente);
-    this.ambienteLiberado = new Ambiente();
-    this.ambienteLiberado.id = ambienteId;
-    this.ambienteLiberado.nome = ambiente;
-
-    this.service.PutLioberarAmbienteBackEnd(this.ambienteLiberado,ambienteId).subscribe(resultado =>{
-      this.snackBar.open(resultado.mensagem,"Ambiente",{
-        duration:2000,
-        horizontalPosition:'center',
-        verticalPosition:'bottom'
-      })
-    },erro => {
-      if(erro.status === '400'){
-       for(const campos in erro.error.errors){
-         if(erro.error.errors.hasOwnProperty(campos)){
-           this.erros.push(erro.error.errors[campos])
-         }
-       }
-      }
-      else{
-        this.erros.push(erro.error)
-      }
-    }
-    )
+    
   }
 }
